ObjectSpawner: Spawns objects either thru API or button press on instance
In its infancy so it looks like complete garbage. Data structures are all over the place.

To use: 

Place the script on an object in the scene.

Add the prefabs you want to spawn and their weights (percent of total num spawned) and the randomization desired. I'm retarded so you have to set randoms to 1.0f if you want original size.